Revert "Use Ginkgo -mod vendor in unit, integration and brats tests"
This reverts commit 7dff9b3d30139523a8ebcf10447bc15ac2a1c83f.

+++ scripts/install_tools.sh
@@ -4,7 +4,9 @@ set -euo pipefail
 cd "$( dirname "${BASH_SOURCE[0]}" )/.."
 source .envrc
 
-go install github.com/onsi/ginkgo/ginkgo
+if [ ! -f .bin/ginkgo ]; then
+  go install github.com/onsi/ginkgo/ginkgo
+fi
 
 export BUNDLE_GEMFILE=cf.Gemfile
 bundle install
